+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/ssl_error_decision.dart
@@ -1,0 +1,12 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// A decision on how to handle an SSL error
+enum SslErrorDecision {
+  /// Prevent the request
+  cancel,
+
+  /// Proceed with the request
+  proceed,
+}