+++ packages/webview_flutter/webview_flutter_wkwebview/darwin/webview_flutter_wkwebview/Sources/webview_flutter_wkwebview/FWFURLAuthenticationChallengeHostApi.m
@@ -4,6 +4,7 @@
 
 #import "./include/webview_flutter_wkwebview/FWFURLAuthenticationChallengeHostApi.h"
 #import "./include/webview_flutter_wkwebview/FWFURLProtectionSpaceHostApi.h"
+#import "./include/webview_flutter_wkwebview/FWFDataConverters.h"
 
 @interface FWFURLAuthenticationChallengeFlutterApiImpl ()
 // BinaryMessenger must be weak to prevent a circular reference with the host API it
@@ -28,6 +29,8 @@
 
 - (void)createWithInstance:(NSURLAuthenticationChallenge *)instance
            protectionSpace:(NSURLProtectionSpace *)protectionSpace
+       sslErrorTypeDataBoxed:(FWFSslErrorTypeDataBox *_Nullable)sslErrorTypeDataBoxed
+          x509CertificateDer:(FlutterStandardTypedData *_Nullable)x509CertificateDer
                 completion:(void (^)(FlutterError *_Nullable))completion {
   if ([self.instanceManager containsInstance:instance]) {
     return;
@@ -36,14 +39,27 @@
   FWFURLProtectionSpaceFlutterApiImpl *protectionSpaceApi =
       [[FWFURLProtectionSpaceFlutterApiImpl alloc] initWithBinaryMessenger:self.binaryMessenger
                                                            instanceManager:self.instanceManager];
-  [protectionSpaceApi createWithInstance:protectionSpace
-                                    host:protectionSpace.host
-                                   realm:protectionSpace.realm
-                    authenticationMethod:protectionSpace.authenticationMethod
-                              completion:^(FlutterError *error) {
-                                NSAssert(!error, @"%@", error);
-                              }];
-
+  
+  if (protectionSpace.authenticationMethod == NSURLAuthenticationMethodServerTrust) {
+    [protectionSpaceApi createWithInstance:protectionSpace
+                         sslErrorTypeDataBoxed:sslErrorTypeDataBoxed
+                       x509CertificateDer:x509CertificateDer
+                                  protocol:protectionSpace.protocol
+                                      host:protectionSpace.host
+                                      port: protectionSpace.port
+                                completion:^(FlutterError *error) {
+      NSAssert(!error, @"%@", error);
+    }];
+  } else {
+    [protectionSpaceApi createWithInstance:protectionSpace
+                                      host:protectionSpace.host
+                                     realm:protectionSpace.realm
+                      authenticationMethod:protectionSpace.authenticationMethod
+                                completion:^(FlutterError *error) {
+                                  NSAssert(!error, @"%@", error);
+                                }];
+  }
+  
   [self.api createWithIdentifier:[self.instanceManager addHostCreatedInstance:instance]
        protectionSpaceIdentifier:[self.instanceManager
                                      identifierWithStrongReferenceForInstance:protectionSpace]